+++ transient/history.el
@@ -1,5 +1,6 @@
 ((magit-commit nil)
  (magit-log
+  ("-n256" "--graph" "--decorate")
   ("-n32" "--stat"))
  (magit-pull nil)
  (magit-push nil))